Add work to task in the past

+++ backend/src/task/task.controller.ts
@@ -189,13 +189,15 @@ export class TaskController {
 
     // Can't delete active or ended tasks
     const task = await this.taskService.getTaskById(taskId);
-    
+    if (!task)
+      throw new NotFoundException();
+
     if (task.category === TaskCategory.ENDED)
       throw new ForbiddenException('Can\'t delete ended task');
 
     // After task is assigned it can be deleted only by assigned user, scrum master or admin 
     if (task.assignedUserId && task.assignedUserId !== token.sid && !token.isAdmin && !await this.storyService.hasUserPermissionForStory(task.storyId, token.sid, UserRole.ScrumMaster))
-      throw new ForbiddenException('Insufficient permissions to delete task');
+      throw new ForbiddenException('Assigned task can only be deleted by assigned user or scrum master');
     
     try {
       await this.taskService.deleteTask(taskId);
@@ -228,7 +230,9 @@ export class TaskController {
 
     // Check if task exists
     const task = await this.taskService.getTaskById(taskId);
-    
+    if (!task)
+      throw new NotFoundException();
+
     const hasOverridePermissions: boolean = token.isAdmin || await this.projectService.hasUserRoleOnProject(projectId, token.sid, UserRole.ScrumMaster);
 
     // User can accept task; admin and scrum master can reassign people
@@ -278,6 +282,8 @@ export class TaskController {
       throw new ForbiddenException("Task isn't part of active sprint");
 
     const task = await this.taskService.getTaskById(taskId);
+    if (!task)
+      throw new NotFoundException();
 
     // User can only accept tasks that were assigned to him
     if (task.assignedUserId !== token.sid)
@@ -313,6 +319,8 @@ export class TaskController {
 
     // Check if someone is already assigned to task or if task even exitsts
     const task = await this.taskService.getTaskById(taskId);
+    if (!task)
+      throw new NotFoundException();
 
     // User can assign only himself, project scrum master can also others
     if (token.sid !== task.assignedUserId && !token.isAdmin && !await this.projectService.hasUserRoleOnProject(await this.taskService.getTaskProjectId(taskId), token.sid, UserRole.ScrumMaster))
@@ -375,6 +383,8 @@ export class TaskController {
 
     // Check if someone is already assigned to task or if task even exitsts
     const task = await this.taskService.getTaskById(taskId);
+    if (!task)
+      throw new NotFoundException();
 
     // User can assign only himself, project scrum master can also others
     if (token.sid !== task.assignedUserId && !token.isAdmin && !await this.taskService.hasUserPermissionForTask(token.sid, taskId, [UserRole.ScrumMaster]))
@@ -403,6 +413,8 @@ export class TaskController {
     
     // Check if someone is already assigned to task or if task even exitsts
     const task = await this.taskService.getTaskById(taskId);
+    if (!task)
+      throw new NotFoundException();
 
     // User can assign only himself, project scrum master can also others
     if (token.sid !== task.assignedUserId && !token.isAdmin && !await this.taskService.hasUserPermissionForTask(token.sid, taskId, [UserRole.ScrumMaster]))
@@ -431,6 +443,8 @@ export class TaskController {
     
     // Check if someone is already assigned to task or if task even exitsts
     const task = await this.taskService.getTaskById(taskId);
+    if (!task)
+      throw new NotFoundException();
 
     // User can assign only himself, project scrum master can also others
     if (token.sid !== task.assignedUserId && !token.isAdmin && !await this.taskService.hasUserPermissionForTask(token.sid, taskId, [UserRole.ScrumMaster]))
@@ -460,11 +474,13 @@ export class TaskController {
       throw new ForbiddenException();
 
     // Check if task part of active sprint
-    if (!await this.taskService.isTaskInActiveSprint(taskId) && !token.isAdmin && !await this.projectService.hasUserRoleOnProject(await this.taskService.getTaskProjectId(taskId), token.sid, UserRole.ScrumMaster))
-      throw new ForbiddenException("Task isn't part of active sprint");
+    if (!await this.taskService.hasBeenTaskInActiveSprint(taskId, date) && !token.isAdmin && !await this.projectService.hasUserRoleOnProject(await this.taskService.getTaskProjectId(taskId), token.sid, UserRole.ScrumMaster))
+      throw new ForbiddenException("Task isn't or wasn't part of active sprint");
 
     // Check if user assigned to task
     const task = await this.taskService.getTaskById(taskId);
+    if (!task)
+      throw new NotFoundException();
 
     // User can assign only himself, project scrum master can also others
     if ((token.sid !== userId || userId !== task.assignedUserId) && !token.isAdmin && !await this.projectService.hasUserRoleOnProject(await this.taskService.getTaskProjectId(taskId), token.sid, UserRole.ScrumMaster))
@@ -502,6 +518,8 @@ export class TaskController {
 
     // Check if user assigned to task
     const task = await this.taskService.getTaskById(taskId);
+    if (!task)
+      throw new NotFoundException();
 
     // User can remove only his time, project scrum master can also others
     if (token.sid !== task.assignedUserId && !token.isAdmin && !await this.projectService.hasUserRoleOnProject(await this.taskService.getTaskProjectId(taskId), token.sid, UserRole.ScrumMaster))
